Piscine Data Science/2 - Data Analyst/requirements.txt
pandas==2.3.3
psycopg2-binary==2.9.11
matplotlib==3.10.7
seaborn==0.13.2
scikit-learn==1.7.2